+++ test/table/test_table.py
@@ -149,6 +149,7 @@ def test_table_with_multiline_cells_and_fixed_row_height(tmp_path):
             for datum in data_row:
                 row.cell(datum)
     assert pdf.pages_count == 2
+
     assert_pdf_equal(
         pdf, HERE / "table_with_multiline_cells_and_fixed_row_height.pdf", tmp_path
     )
@@ -233,9 +234,7 @@ def test_table_with_multiline_cells_and_split_over_3_pages(tmp_path):
                 row.cell(datum)
     assert pdf.pages_count == 4
     assert_pdf_equal(
-        pdf,
-        HERE / "table_with_multiline_cells_and_split_over_3_pages.pdf",
-        tmp_path,
+        pdf, HERE / "table_with_multiline_cells_and_split_over_3_pages.pdf", tmp_path
     )
 
 
@@ -431,6 +430,7 @@ def test_table_with_gutter(tmp_path):
         TABLE_DATA, borders_layout="SINGLE_TOP_LINE", gutter_height=3, gutter_width=3
     ):
         pass
+
     assert_pdf_equal(pdf, HERE / "table_with_gutter.pdf", tmp_path)